refactor(simulation): align service structure to hexagonal conventions
- Move DTOs from dto/ to web/dto/
- Move controllers from infrastructure/adapter/in/web/ to web/controller/
- Update all import references to new web.dto package
- Remove obsolete dto/ folder and broken controller files
