+++ src/main/java/com/renewsim/backend/simulation_service/web/dto/NormalizationStatsDTO.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.simulation_service.dto;
+package com.renewsim.backend.simulation_service.web.dto;
 
 import jakarta.validation.constraints.PositiveOrZero;
