+++ src/main/java/com/renewsim/backend/simulation_service/web/dto/SimulationResponseDTO.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.simulation_service.dto;
+package com.renewsim.backend.simulation_service.web.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
